Refactoring cluster and it's managers to live in the multi_device directory and the cluster manager is now called simply Cluster.
