+++ f5/multi_device/cluster/__init__.py
@@ -15,8 +15,8 @@
 #
 
 from f5.bigip.mixins import DeviceMixin
-from f5.cluster.device_group_manager import DeviceGroupManager as dgm
-from f5.cluster.trusted_peer_manager import TrustedPeerManager as peer_mgr
+from f5.multi_device.cluster.managers import DeviceGroupManager as dgm
+from f5.multi_device.cluster.managers import TrustedPeerManager as peer_mgr
 
 
 class ClusterNotSupported(Exception):
@@ -27,7 +27,7 @@ class RootRemovalNotSupported(Exception):
     pass
 
 
-class ClusterManager(DeviceMixin):
+class Cluster(DeviceMixin):
     '''Manage a cluster of BigIPs.
 
     This is accomplished with REST URI calls only, but some operations are
@@ -40,7 +40,7 @@ class ClusterManager(DeviceMixin):
     def __init__(self, bigips, cluster_name, partition, cluster_type):
         '''Initialize a cluster manager object.
 
-        :param bigips: list -- list of bigips to cluster
+        :param bigips: list -- list of bigip ojects to cluster
         :param cluster_name: str -- name of device service group
         :param partition: str -- partition to deploy cluster to
         :param cluster_type: str -- type of cluster configuration
