+++ f5/multi_device/cluster/managers.py
@@ -51,7 +51,7 @@ class DeviceGroupManager(DeviceMixin):
         '''Create the device service cluster group and add devices to it.'''
 
         self._check_all_devices_in_sync()
-        dg = self.root_device.cm.device_groups.device_group
+        dg = self.root_device.tm.cm.device_groups.device_group
         dg.create(
             name=self.device_group_name,
             partition=self.partition,
@@ -59,7 +59,7 @@ class DeviceGroupManager(DeviceMixin):
         )
         for device in self.devices:
             self._add_device_to_device_group(device)
-            device.sys.config.save()
+            device.tm.sys.config.save()
         self.check_device_group_status()
 
     def check_device_group_status(self):
@@ -77,7 +77,7 @@ class DeviceGroupManager(DeviceMixin):
         '''
 
         self._add_device_to_device_group(device)
-        device.sys.config.save()
+        device.tm.sys.config.save()
         self.devices.append(device)
         self.check_device_group_status()
 
@@ -88,7 +88,7 @@ class DeviceGroupManager(DeviceMixin):
         '''
 
         self._delete_device_from_device_group(device)
-        device.sys.config.save()
+        device.tm.sys.config.save()
         self.devices.remove(device)
         self.check_device_group_status()
 
@@ -108,7 +108,8 @@ class DeviceGroupManager(DeviceMixin):
         self.check_device_group_status()
         for device in self.devices:
             self._delete_device_from_device_group(device)
-            self._ensure_device_active(device)
+            # self._ensure_device_active(device)
+            self.check_device_group_status()
         dg = self._get_device_group(self.root_device)
         dg.delete()
         self.check_devices_active_licensed()
@@ -123,7 +124,7 @@ class DeviceGroupManager(DeviceMixin):
         :returns: tm.cm.device_groups.device_group object
         '''
 
-        return device.cm.device_groups.device_group.load(
+        return device.tm.cm.device_groups.device_group.load(
             name=self.device_group_name, partition=self.partition
         )
 
@@ -191,7 +192,7 @@ class DeviceGroupManager(DeviceMixin):
         '''
 
         self._sync_to_group(device)
-        act = device.cm.devices.device.load(
+        act = device.tm.cm.devices.device.load(
             name=self.get_device_info(device).name,
             partition=self.partition
         )
@@ -206,7 +207,7 @@ class DeviceGroupManager(DeviceMixin):
         '''
 
         config_sync_cmd = 'config-sync to-group %s' % self.device_group_name
-        device.cm.exec_cmd('run', utilCmdArgs=config_sync_cmd)
+        device.tm.cm.exec_cmd('run', utilCmdArgs=config_sync_cmd)
 
     def _ensure_active_standby(self):
         """Ensure cluster is in active standby configuration."""
@@ -260,7 +261,7 @@ class DeviceGroupManager(DeviceMixin):
 
         bigips = []
         for bigip in self.devices:
-            sync_status = bigip.cm.sync_status
+            sync_status = bigip.tm.cm.sync_status
             sync_status.refresh()
             current_status = (sync_status.entries[self.sync_status_entry]
                               ['nestedStats']['entries']['status']
@@ -278,10 +279,133 @@ class DeviceGroupManager(DeviceMixin):
 
         bigips = []
         for bigip in self.devices:
-            act = bigip.cm.devices.device.load(
+            act = bigip.tm.cm.devices.device.load(
                 name=self.get_device_info(bigip).name,
                 partition=self.partition
             )
             if act.failoverState == state:
                 bigips.append(bigip)
         return bigips
+
+
+class TrustedPeerManager(DeviceMixin):
+    '''Manages the trusted peers of a BigIP device.'''
+
+    iapp_actions = {'definition': {'implementation': None, 'presentation': ''}}
+
+    def __init__(self, trust_name, partition):
+        '''Initialize a trusted peer manager object.
+
+        :param trust_name: str -- name of trust to use
+        :param partition: str -- partition to place trusted peers in
+        '''
+
+        self.trust_name = trust_name
+        self.partition = partition
+        self.peer_iapp_prefix = 'cluster_iapp'
+
+    def add_trusted_peers(self, root_bigip, peers):
+        '''Add trusted peers to the root bigip device.
+
+        :param root_bigip: bigip object -- device to add trusted peers to
+        :param peers: list -- bigip objects to add to root device
+        '''
+
+        for peer in peers:
+            self._modify_trusted_peer(peer, self._get_add_peer_cmd, root_bigip)
+
+    def remove_trusted_peers(self, bigip, peer_to_remove=None):
+        '''Remove all trusted peers, unless one is given explicitly.
+
+        :param bigip: bigip object -- bigip to remove peers from
+        :param peer_to_remove: bigip object -- single peer to remove
+        '''
+
+        tds = bigip.tm.cm.trust_domains.trust_domain.load(name=self.trust_name)
+        bigip_info = self.get_device_info(bigip)
+        peer_list = [dv for dv in tds.caDevices if bigip_info.name not in dv]
+        if peer_to_remove:
+            peer_name = self.get_device_info(peer_to_remove)
+            peer_list = [peer for peer in peer_list if peer_name.name in peer]
+        for peer in peer_list:
+            peer_hostname = peer.replace('/%s/' % self.partition, '')
+            self._modify_trusted_peer(
+                peer_hostname, self._get_delete_peer_cmd, bigip
+            )
+
+    def _modify_trusted_peer(
+            self, peer, mod_peer_func, deploy_bigip
+    ):
+        '''Modify a trusted peer device by deploying an iapp.
+
+
+        :param peer: bigip object -- peer to modify
+        :param mod_peer_func: function -- function to call to modify peer
+        :param deploy_bigip: bigip object -- bigip on which to deploy the iapp
+        '''
+
+        iapp_name = '%s_peer' % (self.peer_iapp_prefix)
+        mod_peer_cmd = mod_peer_func(peer)
+        iapp_actions = self.iapp_actions.copy()
+        iapp_actions['definition']['implementation'] = mod_peer_cmd
+        self._deploy_iapp(iapp_name, iapp_actions, deploy_bigip)
+        self._delete_iapp(iapp_name, deploy_bigip)
+
+    def _delete_iapp(self, iapp_name, deploy_bigip):
+        '''Delete an iapp service and template on the root device.
+
+        :param iapp_name: str -- name of iapp
+        :param deploy_bigip: bigip object -- where the iapp will be deleted
+        '''
+
+        iapp = deploy_bigip.tm.sys.applications
+        iapp_service = iapp.services.service.load(
+            name=iapp_name, partition=self.partition
+        )
+        iapp_service.delete()
+        iapp_template = iapp.templates.template.load(
+            name=iapp_name, partition=self.partition
+        )
+        iapp_template.delete()
+
+    def _deploy_iapp(self, iapp_name, actions, deploy_bigip):
+        '''Deploy iapp on the root device.
+
+        :param iapp_name: str -- name of iapp
+        :param actions: dict -- actions definition of iapp sections
+        :param deploy_bigip: bigip object -- bigip object where iapp will be
+            created
+        '''
+
+        tmpl = deploy_bigip.tm.sys.applications.templates.template
+        serv = deploy_bigip.tm.sys.applications.services.service
+        tmpl.create(name=iapp_name, partition=self.partition, actions=actions)
+        serv.create(
+            name=iapp_name, partition=self.partition, template=iapp_name
+        )
+
+    def _get_add_peer_cmd(self, peer):
+        '''Get tmsh command to add a trusted peer.
+
+        :param peer: bigip object -- peer device
+        :returns: str -- tmsh command to add trusted peer
+        '''
+
+        peer_device = self.get_device_info(peer)
+        print('Adding following peer to root: %s' % peer_device.name)
+        username = peer._meta_data['username']
+        password = peer._meta_data['password']
+        return 'tmsh::modify cm trust-domain Root ca-devices add ' \
+            '\\{ %s \\} name %s username %s password %s' % (
+                peer_device.managementIp, peer_device.name, username, password
+            )
+
+    def _get_delete_peer_cmd(self, peer_name):
+        '''Get tmsh command to delete a trusted peer.
+
+        :param peer: bigip object -- peer device
+        :returns: str -- tmsh command to delete trusted peer
+        '''
+
+        return 'tmsh::modify cm trust-domain Root ca-devices delete ' \
+            '\\{ %s \\}' % peer_name
